refactor(styles): Move some styles to come from RowBuilderWidget
