+++ src/qqq/components/widgets/misc/PivotTableValueElement.tsx
@@ -32,7 +32,8 @@ import TextField from "@mui/material/TextField";
 import type {Identifier, XYCoord} from "dnd-core";
 import colors from "qqq/assets/theme/base/colors";
 import FieldAutoComplete from "qqq/components/misc/FieldAutoComplete";
-import {DragItemTypes, fieldAutoCompleteTextFieldSX, getSelectedFieldForAutoComplete, xIconButtonSX} from "qqq/components/widgets/misc/PivotTableSetupWidget";
+import {DragItemTypes, fieldAutoCompleteTextFieldSX, getSelectedFieldForAutoComplete} from "qqq/components/widgets/misc/PivotTableSetupWidget";
+import {xIconButtonSX} from "qqq/components/widgets/misc/RowBuilderWidget";
 import {functionsPerFieldType, PivotTableDefinition, pivotTableFunctionLabels, PivotTableValue} from "qqq/models/misc/PivotTableDefinitionModels";
 import React, {FC, useReducer, useRef, useState} from "react";
 import {useDrag, useDrop} from "react-dnd";
